+++ contracts/services/ICrunaService.sol
@@ -17,7 +17,7 @@ interface ICrunaService is IERC7656Contract, IVersioned {
   /**
    * @notice Initialize the plugin. It must be implemented, but can do nothing is no init is needed.
    */
-  function init() external;
+  function init(bytes memory data) external;
 
   /**
    * @notice Called by the manager to know if the plugin is an ERC721 account